requirements.txt
fastapi
uvicorn
python-jose[cryptography]
sqlalchemy
passlib
python-multipart
psycopg2
alembic
aiofiles
jinja2